backend/requirements.txt
django==1.7
django-extensions==1.4.0
django-getenv==1.3.1
django-dotenv==1.2
djangorestframework==2.4.2
django-grappelli==2.6.1
gunicorn==19.1.1
psycopg2==2.5.4
raven==5.0.0
